Adding multi-threading support. Part 3.

+++ src/mb4d_gui/work_queue.hpp
@@ -8,7 +8,7 @@ class WorkQueue {
 
     WorkQueueItem* workQueueItems;
   public:
-    void initNewQueueItem(double, double, double, double);
+    void initNewQueueItem(unsigned int, unsigned int, unsigned int, unsigned int);
     void push_item(WorkQueueItem*);
 
     void clearQueue(void);
